@@ -1,27 +1,38 @@
 {
   "name": "functions",
   "description": "Cloud Functions for Firebase",
+  "main": "dist/",
   "scripts": {
+    "build": "tslint --project tsconfig.json && webpack",
     "lint": "eslint .",
     "serve": "firebase serve --only functions",
     "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
+    "logs": "firebase functions:log",
+    "emulator": "firebase emulators:start && webpack --watch"
   },
   "engines": {
-    "node": "8"
+    "node": "10"
   },
   "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "firebase": "^7.8.1",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.3.0",
-    "uuid-token-generator": "^1.0.0"
+    "tslint": "^6.0.0",
+    "typescript": "^3.8.2",
+    "uuid-token-generator": "^1.0.0",
+    "webpack": "^4.41.6"
   },
   "devDependencies": {
     "eslint": "^5.12.0",
     "eslint-plugin-promise": "^4.0.1",
-    "firebase-functions-test": "^0.1.6"
+    "firebase-functions-test": "^0.1.6",
+    "ts-loader": "^6.2.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-node-externals": "^1.7.2"
   },
   "private": true
 }
