@@ -12,12 +12,14 @@ module.exports = {
         '_helpers/*': './helpers/*',
         '_assets/*': './assets/*',
         '_components/*': './screens/components/*',
+        _types: './types/index.ts',
         _navigation: './navigationService.js',
         _typography: './theme/typography.js',
         _palette: './theme/palette',
-        // _metrics: './theme/metrics',
+        _metrics: './theme/metrics',
         _globals: './theme/globals',
         _zIndex: './theme/zIndex',
+        _rootReducer: './reducers/index.ts',
       },
     },
   ],
