@@ -1,29 +1,29 @@
-import React from "react";
-import { PersistGate } from "redux-persist/integration/react";
-import { Provider } from "react-redux";
-import * as Font from "expo-font";
-import { View } from "react-native";
-import { StyleProvider } from "native-base";
+import React from 'react';
+import { PersistGate } from 'redux-persist/integration/react';
+import { Provider } from 'react-redux';
+import * as Font from 'expo-font';
+import { View } from 'react-native';
+import { StyleProvider } from 'native-base';
 
-import { SafeAreaProvider } from "react-native-safe-area-view";
-import * as navigationService from "./navigationService";
-import getTheme from "./native-base-theme/components";
-import AppContainer from "./screens/AppContainer";
-import { store, persistor } from "./store";
-import reactotron from "reactotron-react-native";
+import { SafeAreaProvider } from 'react-native-safe-area-view';
+import * as navigationService from './navigationService';
+import getTheme from './native-base-theme/components';
+import AppContainer from './screens/AppContainer';
+import { store, persistor } from './store';
+import reactotron from 'reactotron-react-native';
 
 console.disableYellowBox = true;
 
 export default class App extends React.Component {
   state = {
-    fontLoaded: false
+    fontLoaded: false,
   };
 
   async componentDidMount() {
     if (!this.state.fontLoaded) {
       await Font.loadAsync({
-        prompt: require("./assets/fonts/Prompt.ttf"),
-        rubik: require("./assets/fonts/rubik.ttf")
+        prompt: require('./assets/fonts/Prompt.ttf'),
+        rubik: require('./assets/fonts/rubik.ttf'),
       });
       this.setState({ fontLoaded: true });
     }
@@ -38,7 +38,9 @@ export default class App extends React.Component {
               <StyleProvider style={getTheme()}>
                 <AppContainer
                   ref={navigationRef => {
-                    navigationService.setTopLevelNavigator(navigationRef);
+                    navigationService.setTopLevelNavigator(
+                      navigationRef,
+                    );
                   }}
                 />
               </StyleProvider>
