@@ -2,6 +2,7 @@ import database from '_apis/database';
 import * as types from '../app';
 import { navigate } from '_navigation';
 import reactotron from 'reactotron-react-native';
+import { AppThunk } from '_types';
 
 export const createUser = (
   userData: {
@@ -31,7 +32,7 @@ export const signIn = (
   username: string,
   password: string,
   onSignInFailed: Function,
-) => async dispatch => {
+): AppThunk => async dispatch => {
   try {
     const response = await database.post('signIn', {
       email: username,
@@ -47,11 +48,10 @@ export const signIn = (
       },
     });
   } catch (error) {
-    const { data } = error.response;
     onSignInFailed('Invalid password or email');
   }
 };
-export const signOut = (onFailed: Function) => async (
+export const signOut = (onFailed: Function): AppThunk => async (
   dispatch,
   getState,
 ) => {
@@ -66,8 +66,9 @@ export const signOut = (onFailed: Function) => async (
 export const checkAuth = (
   onAuthSucces: Function,
   onAuthFailed: Function,
-) => async (dispatch, getState) => {
+): AppThunk => async (dispatch, getState) => {
   const state = getState();
+  console.log('Check');
   try {
     const { token, uid } = state.app.user;
     const response = await database.post('checkAuth', { uid, token });
@@ -77,6 +78,7 @@ export const checkAuth = (
     dispatch({
       type: types.CHECK_AUTH,
       payload: {
+        ...data,
         token: newToken,
       },
     });
