@@ -24,6 +24,7 @@ class Loading extends React.Component {
   constructor(props) {
     super(props);
     this.unsubscribe = NetInfo.addEventListener(state => {
+      console.log(state.isConnected);
       this.setState({ isConnected: state.isConnected });
     });
   }
@@ -38,14 +39,14 @@ class Loading extends React.Component {
     this.navigateScreens();
   }
 
-  componentWillUnmount() {
-    this.unsubscribe();
-  }
+  // componentWillUnmount() {
+  //   this.unsubscribe();
+  // }
 
   navigateScreens = () => {
     const { isConnected } = this.state;
     const { signedIn } = this.props;
-    console.log(signedIn);
+    console.log(`Connected  ${isConnected}`);
     if (signedIn) {
       if (isConnected) {
         this.props.checkAuth(
@@ -53,7 +54,7 @@ class Loading extends React.Component {
             navigate('latest');
           },
           () => {
-            navigate('signIn');
+            navigate('login');
           },
         );
       } else navigate('latest');
@@ -106,4 +107,4 @@ const mapStateToProps = state => {
   };
 };
 
-export default connect(mapStateToProps)(Loading);
+export default connect(mapStateToProps, { checkAuth })(Loading);
