@@ -7,13 +7,15 @@ import typography from '_typography';
 import { TouchableWithoutFeedback } from 'react-native-gesture-handler';
 
 interface Props {
-  style: CSSProperties;
-  secondary: boolean;
+  style?: CSSProperties;
+  buttonStyle?: CSSProperties;
+  secondary?: boolean;
   onPress(event: GestureResponderEvent): void;
+  title: string;
 }
 
 const Button = (props: Props) => {
-  const { style, secondary, onPress } = props;
+  const { style, secondary, onPress, title, buttonStyle } = props;
   const backgroundColor = secondary
     ? palette.secondary
     : palette.primary;
@@ -21,7 +23,9 @@ const Button = (props: Props) => {
   return (
     <TouchableWithoutFeedback onPress={onPress}>
       <View style={[styles.container, { backgroundColor }, style]}>
-        <Text style={[styles.text, { color }]}>Sign in</Text>
+        <Text style={[styles.text, { color }, buttonStyle]}>
+          {title}
+        </Text>
       </View>
     </TouchableWithoutFeedback>
   );
