@@ -1,19 +1,20 @@
-import { PixelRatio } from "react-native";
+import { PixelRatio } from 'react-native';
 
 const scaleFont = size => size * PixelRatio.getFontScale();
 
 export default {
   fonts: {
-    primary: "prompt"
+    primary: 'rubik',
   },
   fontWeight: {
-    bold: "bold",
-    regular: "400"
+    bold: 'bold',
+    regular: '400',
   },
   fontSize: {
+    verySmall: scaleFont(13),
     small: scaleFont(15),
     normal: scaleFont(18),
     medium: scaleFont(22),
-    big: scaleFont(28)
-  }
+    big: scaleFont(28),
+  },
 };
